[dag] rename reliable_broadcast.rs to rb_handler.rs
