+++ consensus/src/dag/mod.rs
@@ -9,7 +9,7 @@ mod dag_handler;
 mod dag_network;
 mod dag_store;
 mod order_rule;
-mod reliable_broadcast;
+mod rb_handler;
 mod storage;
 #[cfg(test)]
 mod tests;
