+++ consensus/src/dag/tests/mod.rs
@@ -7,5 +7,5 @@ mod dag_test;
 mod fetcher_test;
 mod helpers;
 mod order_rule_tests;
-mod reliable_broadcast_tests;
+mod rb_handler_tests;
 mod types_test;
